pre sale option is remove from invoice creator page

+++ stock/forms.py
@@ -19,7 +19,7 @@ class ProductMaterialForm(forms.ModelForm):
             }
         ))
 
-    quantity = forms.IntegerField(
+    quantity = forms.FloatField(
         label='Enter Quatity of Material',
         widget=forms.NumberInput(
             attrs={
